修改：修复了一个bug，该bug导致接口参数只能用特定类型来接受，现在可以用任意类型了

+++ src/main/java/cn/ayl/intf/Organize.java
@@ -21,7 +21,7 @@ public interface Organize extends IMicroService {
 
     @Method(desc = "获取用户", auth = true)
     JsonObject getUser(
-            @Param("用户id") Integer userId
+            @Param("用户id") Long userId
     );
 
 }
